[TESTS] Updates utilities
