+++ tests/api/__init__.py
@@ -1,0 +1,4 @@
+#
+# Copyright(c) 2019 Intel Corporation
+# SPDX-License-Identifier: BSD-3-Clause-Clear
+#
